Download test microbiome data

Test microbiome data in phyloseq format.

get_microbiome_data(disease, study)

Arguments

disease

Disease of interest as shown in list_microbiome_data()

study

Name of the study as shown in list_microbiome_data()

Value

phyloseq-class object.

Details

You can download few example datasets in phyloseq format from Duvallet et al 2017 https://www.nature.com/articles/s41467-017-01973-8.pdf?origin=ppub. The source file for these data is the microbiomedatarepo https://github.com/microsud/microbiomedatarepo.

Examples

if (FALSE) {
# Example data
library(microbiome)
library(microbiomeUtilities)
list_microbiome_data()
ps1 <- get_microbiome_data(disease = "CDI", "Schubert2014_CDI")
print_ps(ps1)
}