Join otu_table and tax_table to Tibble

Utility to join otu_table and tax_table to tibble.

join_otu_tax(x, column_id = "OTUID")

Arguments

x
phyloseq-class object

column_id
Provide name for the column which will hold the rownames of slot.

Value

A tibble

Examples

library(microbiomeutilities)
data("zackular2014")
p0 <- zackular2014
otu_tax <- join_otu_tax(p0,column_id = "OTUID")
head(otu_tax)

#> # A tibble: 6 x 96
#>   OTUID Domain Phylum Class Order Family Genus Species `Adenoma10-2757`
#>   <chr> <chr>  <chr>  <chr> <chr> <chr>  <chr> <chr>              <int>
#> 1 d__d~ k__Ba~ p__Ba~ c__B~ o__B~ f__Ba~ g__B~ s__                    0
#> 2 d__d~ k__Ba~ p__Ba~ c__B~ o__B~ f__Ba~ g__B~ s__                    0
#> 3 d__d~ k__Ba~ p__Fi~ c__C~ o__C~ f__Ru~ g__   s__                    0
#> 4 d__d~ k__Ba~ p__Ba~ c__B~ o__B~ f__Ba~ g__B~ s__                    0
#> 5 d__d~ k__Ba~ p__Fi~ c__C~ o__C~ f__La~ g__B~ s__                    1
#> 6 d__d~ k__Ba~ p__Ba~ c__B~ o__B~ f__Ba~ g__B~ s__                    0
#> # ... with 87 more variables: `Adenoma11-2775` <int>, `Adenoma13-2803` <int>,
#> #   `Adenoma14-2817` <int>, `Adenoma15-2847` <int>, `Adenoma16-2873` <int>,
#> #   `Adenoma17-2937` <int>, `Adenoma18-2967` <int>, `Adenoma19-2971` <int>,
#> #   `Adenoma20-2995` <int>, `Adenoma21-3129` <int>, `Adenoma2-2113` <int>,
#> #   `Adenoma22-3133` <int>, `Adenoma23-3141` <int>, `Adenoma24-3217` <int>,
#> #   `Adenoma25-3273` <int>, `Adenoma26-3373` <int>, `Adenoma27-3379` <int>,
#> #   `Adenoma28-3391` <int>, `Adenoma29-3399` <int>, `Adenoma30-3509` <int>,
#> #   `Adenoma3-2123` <int>, `Adenoma4-2143` <int>, `Adenoma5-2243` <int>,
#> #   `Adenoma6-2453` <int>, `Adenoma7-2475` <int>, `Adenoma8-2555` <int>,
#> #   `Adenoma9-2717` <int>, `Cancer10-2567` <int>, `Cancer11-2573` <int>,
#> #   `Cancer12-2575` <int>, `Cancer1-2355` <int>, `Cancer13-2579` <int>,
#> #   `Cancer14-2611` <int>, `Cancer15-2613` <int>, `Cancer16-2619` <int>,
#> #   `Cancer17-2625` <int>, `Cancer18-2629` <int>, `Cancer19-2633` <int>,
#> #   `Cancer20-2667` <int>, `Cancer21-2671` <int>, `Cancer22-2703` <int>,
#> #   `Cancer2-2455` <int>, `Cancer23-2705` <int>, `Cancer24-2727` <int>,
#> #   `Cancer25-2735` <int>, `Cancer26-2763` <int>, `Cancer27-2771` <int>,
#> #   `Cancer28-2777` <int>, `Cancer29-2801` <int>, `Cancer30-2863` <int>,
#> #   `Cancer3-2513` <int>, `Cancer4-2521` <int>, `Cancer5-2523` <int>,
#> #   `Cancer6-2535` <int>, `Cancer7-2537` <int>, `Cancer8-2543` <int>,
#> #   `Cancer9-2547` <int>, `Healthy10-3087` <int>, `Healthy11-3123` <int>,
#> #   `Healthy1-2027` <int>, `Healthy12-3211` <int>, `Healthy13-3241` <int>,
#> #   `Healthy14-3245` <int>, `Healthy15-3247` <int>, `Healthy16-3251` <int>,
#> #   `Healthy17-3257` <int>, `Healthy18-3293` <int>, `Healthy19-3339` <int>,
#> #   `Healthy20-3341` <int>, `Healthy21-3347` <int>, `Healthy22-3363` <int>,
#> #   `Healthy2-2779` <int>, `Healthy23-3367` <int>, `Healthy24-3403` <int>,
#> #   `Healthy25-3405` <int>, `Healthy26-3407` <int>, `Healthy27-3439` <int>,
#> #   `Healthy28-3499` <int>, `Healthy29-3529` <int>, `Healthy30-3291-66` <int>,
#> #   `Healthy3-3021` <int>, `Healthy4-3023` <int>, `Healthy5-3043` <int>,
#> #   `Healthy6-3061` <int>, `Healthy7-3063` <int>, `Healthy8-3073` <int>,
#> #   `Healthy9-3075` <int>