Microbiome analysis pipeline

Function microbiome_pipeline generates an HTML report with preliminary QC, Alpha Diversity, Ordination and Composition analysis of OTU tables. This function save all intermediate files incuding figures and phyloseq objects in user specified directory.

microbiome_pipeline(otufile, mapping, taxonomy, treefilename, type,
  work_dir, out_dir, VariableA, VariableB, UnConstOrd, heatmap,
  filterCount, filterPrev, col.palette, filterpseq, samsize, projectname,
  author)

Arguments

otufile
biom object, otu_table in csv format or mothur shared files.

mapping
Metadata variable to check for groups based sequencing effort csv format.

taxonomy
NULL or csv fomatted file.

treefilename
For phylogenetic based diversity analysis "*.tre" file

type
"biom", "mothur", "simple" simple is for *.csv file.

work_dir
Working directory where input files are stored.

out_dir
Output directory where all outputs are to be stored.

VariableA
Main variable of interest.

VariableB
Secondary variable of interest.

UnConstOrd
If unconstrained ordination to be plotted TRUE or FALSE.

heatmap
Heatmap or not option is TRUE or FALSE.

filterCount
Filter OTUs below this count number.

filterPrev
Filter OTUs not detected in less than this percent of samples.

col.palette
'Spectral', 'Set2', 'Paired' use any of the RColorbrewer palette depending on number of groups you have in VaraibleA.

filterpseq
TRUE or FALSE.

samsize
Number of reads to rarefy your data and save the rarefied phyobject.

projectname
Name of the project.

author
Name of the author/investigator.

Value

A HTML report with graphs and data stats which can serve as guide for actual analysis.

Examples

# NOT RUN {
library(microbiomeutilities)
library(microbiome)
microbiome_pipeline(otufile = "my.biom",
                   mapping = "mymap.csv",
                   taxonomy = NULL,
                   treefilename = "myTree.tre",
                   type = "biom",
                   work_dir = "F:/path/my/input/filefolder",
                   out_dir = "F:/path/to/save/my/files/folder",
                   VariableA = "MC_type1",
                   VariableB = "Region",
                   UnConstOrd = TRUE,
                   heatmap = TRUE,
                   filterCount = 4,
                   filterPrev = 0.01,
                   col.palette = "Paired",
                   filterpseq = TRUE,
                   samsize = NA,
                   projectname = "Mock",
                   author = "Sudarshan")

# }