Rarefaction curves for alpha diversity indices

Calculates alpha diversity idenx at varying sampling units (sequencing depth).

plot_alpha_rcurve(
  x,
  index = "observed",
  subsamples = c(100, 1000, 2000, 3000, 4000, 5000),
  lower.conf = 0.025,
  upper.conf = 0.975,
  group = NULL,
  linetype.main = 1,
  line.opacity.main = 0.5,
  linetype.type = 2,
  line.opacity.type = 0.25,
  type = "CI"
)

Arguments

x
phyloseq-class object

index
Default:: "observed",

subsamples
Default: c(100,1000, 2000, 3000, 4000, 5000)

lower.conf
Default: 0.025. If type=CI

upper.conf
Default: 0.975.

group
Default: NULL

linetype.main
For ggplot line type for line by group. Default: 1

line.opacity.main
For ggplot alpha to determine opacity for line by group. Default: 0.5

linetype.type
For ggplot line type for line CI or SD. Default: 2

line.opacity.type
For ggplot line type to determine opacity for line CI or SD. Default: 0.25

type
Either CI (confidence interval) or SD (Standard deviation) Default: CI

Examples

if (FALSE) {
library(microbiomeutilities)
data("zackular2014")
p0 <- zackular2014
p <- plot_alpha_rcurve(p0, index="observed", 
lower.conf = 0.025, upper.conf = 0.975, 
group="DiseaseState") + 
scale_color_brewer(palette = "Paired") + 
scale_fill_brewer(palette = "Paired")
print(p )
}