Taxonomic Composition Plot

Plot taxon abundance for samples. It is a legacy function from microbiome.

plot_taxa_composition(
  x,
  sample.sort = NULL,
  taxonomic.level = "Phylum",
  transform = "compositional",
  otu.sort = NULL,
  palette = brewer.pal(12, "Paired"),
  x.label = "sample",
  plot.type = "barplot",
  average_by = NULL,
  verbose = FALSE,
  mar = c(5, 12, 1, 1),
  ...
)

Arguments

x

phyloseq-class object

sample.sort

Order samples. Various criteria are available:

NULL or 'none': No sorting

A single character string: indicate the metadata field to be used for ordering

A character vector: sample IDs indicating the sample ordering.

'neatmap' Order samples based on the neatmap approach. See neatsort. By default, 'NMDS' method with 'bray' distance is used. For other options, arrange the samples manually with the function.

taxonomic.level

Merge the OTUs (for phyloseq object) into a higher taxonomic level. This has to be one from colnames(tax_table(x)).

transform

Data transform to be used in plotting (but not in sample/taxon ordering). The options are 'Z-OTU', 'Z-Sample', 'log10' and 'compositional'. See the transform function.

otu.sort

Order taxa. Same options as for the sample.sort argument but instead of metadata, taxonomic table is used. Also possible to sort by 'abundance'.

palette

The number and palette RColorBrewer has to be specified e.g brewer.pal(12, "Paired").

x.label

Specify how to label the x axis. This should be one of the variables in sample_variables(x).

plot.type

Plot type: 'barplot' or 'lineplot'.

average_by

Variable to group.

verbose

verbose.

mar

Figure margins.

...

Arguments to be passed (for neatsort function)

Value

A ggplot plot object.

Examples

if (FALSE) {
# Example data
library(microbiome)
library(microbiomeutilities)
data("biogeogut")
pseq <- biogeogut
plot_taxa_composition(pseq, taxonomic.level = "Phylum")
}