Create table for ternary plot OTU

Create a table for ternary plot ggtern package.

prep_tern_otu(x, abund.thres = 1e-04, prev.thres = 0.1, group = NULL)

Arguments

x
phyloseq-class object.

abund.thres
= 0.0001 check microbiome package core function. remove taxa that are detected at 0.0001 in less than prev.thres of samples.

prev.thres
= 0.1 check microbiome package core function.

group
Grouping variable to compare, for this plot there has to be three groups in the data

Value

tibble object.

Details

Plots the mean relative abundance of taxa in 3 groups being compared.

Examples

# library(microbiome)
# library(microbiomeutilities)
# library(dplyr)
# data("zackular2014")
# p0 <- zackular2014
# prep_tern_otu(p0, group="DiseaseState",
# abund.thres=0.0001, prev.thres=0.25)