Create table for Ternary plot

Create a table for ternary plot ggtern R package.

prep_ternary(
  x,
  abund.thres = 1e-04,
  prev.thres = 0.1,
  group = NULL,
  level = "lowest"
)

Arguments

x
phyloseq-class object

abund.thres
= 0.0001 check microbiome package core function remove taxa that are dectected at 0.0001 in less than prev.thres of samples

prev.thres
= 0.1 check microbiome package core function

group
Grouping variable to compare, for this plot there has to be three groups in the data

level
= "Genus" Taxonomic level. If OTU/ASV level specify="lowest" Does not support phylum level aggregation

Value

Tibble object.

Details

Plots the mean relative abundance of taxa in 3 groups being compared.

Examples

library(microbiome)
library(microbiomeutilities)
library(dplyr)
data("zackular2014")
p0 <- zackular2014
prep_ternary(p0, group = "DiseaseState", abund.thres = 0.0001, level = "Genus", prev.thres = 0.25)

#> An additonal column Sam_rep with sample names is created for reference purpose

#> # A tibble: 73 x 11
#> # Groups:   unique [73]
#>    unique      CRC      H  nonCRC Domain  Phylum  Class Order Family Genus OTUID
#>    <chr>     <dbl>  <dbl>   <dbl> <chr>   <chr>   <chr> <chr> <chr>  <chr> <chr>
#>  1 Acetan~ 0.00995 0.0135 0.00858 Bacter~ "Firmi~ "Clo~ "Clo~ "Rumi~ "Ace~ Acet~
#>  2 Acidam~ 0.115   0.164  0.0846  Bacter~ "Firmi~ "Neg~ "Sel~ "Acid~ "Aci~ Acid~
#>  3 Actino~ 0.0169  0.0338 0.0218  Bacter~ "Actin~ "Act~ "Act~ "Acti~ "Act~ Acti~
#>  4 Akkerm~ 1.26    1.05   1.53    Bacter~ "Verru~ "Ver~ "Ver~ "Verr~ "Akk~ Akke~
#>  5 Alisti~ 0.929   0.894  0.654   Bacter~ "Bacte~ "Bac~ "Bac~ "Rike~ "Ali~ Alis~
#>  6 Anaero~ 0.00693 0.0132 0.00458 Bacter~ "Firmi~ "Clo~ "Clo~ "Clos~ "Ana~ Anae~
#>  7 Anaero~ 0.721   0.795  0.840   Bacter~ "Firmi~ "Clo~ "Clo~ "Lach~ "Ana~ Anae~
#>  8 Anaero~ 0.00364 0.0127 0.00499 Bacter~ "Firmi~ "Clo~ "Clo~ "Clos~ "Ana~ Anae~
#>  9 Bacter~ 0.0395  0.0724 0.0761  Bacter~ ""      ""    ""    ""     ""    Bact~
#> 10 Bacter~ 0.0171  0.0271 0.137   Bacter~ "Actin~ "Act~ "Cor~ "Cori~ ""    Bact~
#> # ... with 63 more rows